requirements.txt
gspread==5.11.2
google-auth==2.27.0
google-auth-oauthlib==1.2.0
selenium==4.16.0
requests==2.31.0
python-dateutil==2.8.2
PyYAML==6.0.1
